pyproject.toml
[tool.poetry]
name = "cv-analyser"
version = "0.1.0"
description = ""
authors = ["Robert Siipola <robert@three.consulting>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
llvmlite = "^0.39.1"
openai = "^0.27.0"
pandas = "^1.5.3"
python-dotenv = "^1.0.0"
gradio = "^3.20.0"
torch = "^1.11.0"
chromadb = "^0.3.10"
pydantic = "^1.10.5"
pyarrow = "^11.0.0"
loguru = "^0.6.0"
setuptools = "^67.4.0"
transformers = "^4.26.1"
umap-learn = "^0.5.3"
kneed = "^0.8.2"
plotly = "^5.13.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ruff = "^0.0.254"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonr^.api"
